pyproject.toml
[project]
name = "dynamic-data-masking"
version = "0.1.0"
description = "Dynamic masking engine for sensitive and personal data in documents."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.4.0",
    "fastapi>=0.115.0",
    "presidio-analyzer>=2.2.0",
    "pydantic-settings>=2.7.0",
    "psycopg[binary]>=3.2.0",
    "pymupdf>=1.25.0",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.32.0",
    "en-core-web-sm",
    "prometheus-client>=0.21.0",
    "python-docx>=1.2.0",
    "alembic>=1.18.4",
]

[project.scripts]
ddm-api = "apps.api.main:run"
ddm-worker = "apps.worker.main:run"

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apps", "src/ddm_engine", "src/dynamic_data_masking"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
